backend/requirements.txt
moviepy>=1.0.3
Pillow>=10.0.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
psutil>=5.9.5
bleach>=6.1.0

openai>=1.40.0
elevenlabs==0.2.26
replicate>=0.20.0

Flask>=3.0.0
Flask-CORS>=4.0.0

gunicorn>=21.2.0

decorator>=4.4.2
proglog>=0.1.10
tqdm>=4.66.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9